ported the h2stdinc.h stuff from hhexen. there is much to be done, yet, especially the configure stuff needs lovin...
git-svn-id: svn://svn.code.sf.net/p/hhexen/code/trunk/hheretic@361 989aa981-fe4f-0410-aa6a-c19d413bcdcd

+++ base/i_sound.c
@@ -1,7 +1,6 @@
 // $Id$
 
-
-#include <stdio.h>
+#include "h2stdinc.h"
 #include <math.h>       // pow()
 #include <pthread.h>
 #include "doomdef.h"
@@ -125,7 +124,7 @@ typedef struct
 {
     short a;            // always 3
     short freq;         // always 11025
-    long length;        // sample length
+    int32_t length;     // sample length
     unsigned char firstSample;
 } Sample;
