+++ base/r_main.c
@@ -1,6 +1,6 @@
 // R_main.c
 
-#include <stdlib.h>
+#include "h2stdinc.h"
 #include <math.h>
 #include "doomdef.h"
 #include "r_local.h"
@@ -58,7 +58,7 @@ int			viewangletox[FINEANGLES/2];
 angle_t		xtoviewangle[SCREENWIDTH+1];
 
 // the finetangentgent[angle+FINEANGLES/4] table holds the fixed_t tangent
-// values for view angles, ranging from MININT to 0 to MAXINT.
+// values for view angles, ranging from H2MININT to 0 to H2MAXINT.
 // fixed_t		finetangent[FINEANGLES/2];
 
 // fixed_t		finesine[5*FINEANGLES/4];
@@ -317,7 +317,7 @@ void R_InitPointToAngle (void)
 // now getting from tables.c
 #if 0
 	int	i;
-	long	t;
+	int	t;	/* int32_t */
 	float	f;
 //
 // slope (tangent) to angle lookup
